16.6430%

+++ include/functions.h
@@ -9,7 +9,7 @@ void *malloc(s32 size);
 s32 rand(void);
 void *memcpy(void *dest, void *src, u32 n);
 
-void *getPointerTableFile(s32 pointerTableIndex, s32 fileIndex, s32 arg2, s32 arg3);
+void *getPointerTableFile(s32 pointerTableIndex, s32 fileIndex, s32 arg2, u8 arg3);
 void loadExits(s32 map);
 void setFlag(s16 flagIndex, u8 newValue, u8 flagType); // TODO: Can we use the FlagTypes enum?
 u8 isFlagSet(s16 flagIndex, u8 flagType); // TODO: Can we use the FlagTypes enum?
